* remove lit component * disable search component * updated @getcanary to 1.0.0-rc release * updated canary-search component based on https://github.com/fastrepl/canary/issues/109 * add loading skeleton
