+++ next.config.mjs
@@ -1,4 +1,3 @@
-import createLitWrapper from "@lit-labs/nextjs"
 import createMDXPlugin from "@next/mdx"
 import { rehypePlugins, remarkPlugins } from "@renoun/mdx-plugins"
 
@@ -9,8 +8,6 @@ const withMDX = createMDXPlugin({
     rehypePlugins,
   },
 })
-
-const withLit = createLitWrapper({ addDeclarativeShadowDomPolyfill: true })
 
 export default withMDX({
   output: "export",
